+++ src/widgets/history/api/fetchPlayerHistory.ts
@@ -1,4 +1,4 @@
-import type { PlayerHistoryResponse } from '../model/types'
+import type { PlayerHistoryResponse } from '../model/type'
 
 export async function fetchPlayerHistory(playerId: number) {
   const res = await fetch(`/api/players/${playerId}/history`)
