+++ src/widgets/history/model/type.ts
@@ -28,3 +28,10 @@ export interface PlayerHistoryTabProps {
   playerId: number
   position: Position
 }
+
+export interface PlayerHistoryResponse {
+  data: {
+    history: SeasonHistory[]
+    growthSummary: GrowthSummary
+  }
+}
